0.5.6 update
